increase timeout

+++ src/mutate.py
@@ -4,7 +4,7 @@ mutation_rate = 0.15
 
 
 def mutate(individual):
-    for i in range(individual.coloring.size / 20):
+    for i in range(int(individual.coloring.size / 20)):
         index1 = random.randint(0, individual.coloring.size - 1)
         index2 = random.randint(0, individual.coloring.size - 1)
         tmp = individual.coloring[index1]
